ocr/requirements.txt
pymupdf 
python-docx 
pillow 
pytesseract
transformers
